HelloFxml2
This is a copy of the original HelloFXML example which tests scala-suffix-maven-plugin -
a Maven plugin which fixes Scala dependencies incompatible with Java 9+.
The app uses a Scala library, wire-signals_2.13:0.3.1. Without the plugin, using this library
would fail because in Java 9+ the suffix _2.13 is invalid. The plugin adds a line to
the manifest file, "Automatic-Module-Name: wire-signals" which fixes that.
(The current version wire-signals, 0.4.2, already has that line, so I use an old version here)

Look here for details: https://github.com/makingthematrix/scala-suffix